Reorganised fixtures to make them reuseable
