+++ tests/ethereum/polygon_forked/generic-router/README.md
@@ -1,0 +1,5 @@
+Generic router specific mainnet fork tests
+
+- Test generic routing execution model
+- Run different strategy loops against a mainnet fork to see generic router
+  works as a part of the strategy decision cycle
